requirements.txt
flask==3.1.2
psycopg2-binary==2.9.11
